requirements.txt
mysql-connector-python==9.0.0
pyTelegramBotAPI==4.22.1
python-decouple==3.8